refactor: rename image model to file model
