+++ models/supplier.js
@@ -3,7 +3,7 @@ const validators = require('./helpers/validators');
 module.exports = {
   _image: {
     type: 'ObjectId',
-    ref: 'Image',
+    ref: 'File',
   },
   name: {
     type: 'String',
